+++ src/uwt.ml
@@ -1650,13 +1650,13 @@ module Fs = struct
   let openfile ?(perm=0o644) ~mode fln =
     Req.ql ~typ ~name:"open" ~param:fln ~f:(openfile fln mode perm)
 
-  external read:
-    file -> 'a -> int -> int ->
+  external iread:
+    file -> 'a -> int -> int -> int64 ->
     loop -> Req.t -> int_cb ->
     Int_result.unit =
     "uwt_fs_read_byte" "uwt_fs_read_native"
 
-  let read ?(pos=0) ?len t ~buf ~dim =
+  let iread ?(pos=0) ?len ~fd_offset t ~buf ~dim =
     let len =
       match len with
       | None -> dim - pos
@@ -1665,23 +1665,35 @@ module Fs = struct
     if pos < 0 || len < 0 || pos > dim - len then
       Lwt.fail (Invalid_argument "Uwt.Fs.read")
     else
-      Req.qli ~typ ~name:"read" ~param ~f:(read t buf pos len)
+      Req.qli ~typ ~name:"read" ~param ~f:(iread t buf pos len fd_offset)
 
   let read_ba ?pos ?len t ~(buf:buf) =
     let dim = Bigarray.Array1.dim buf in
-    read ?pos ?len ~dim ~buf t
+    iread ~fd_offset:Int64.minus_one ?pos ?len ~dim ~buf t
 
   let read ?pos ?len t ~buf =
     let dim = Bytes.length buf in
-    read ?pos ?len ~dim ~buf t
+    iread ~fd_offset:Int64.minus_one ?pos ?len ~dim ~buf t
 
-  external write:
-    file -> 'a -> int -> int ->
-    loop -> Req.t -> int_cb ->
-    Int_result.unit =
-    "uwt_fs_write_byte" "uwt_fs_write_native"
+  let pread_ba ?pos ?len t ~fd_offset ~(buf:buf) =
+    if Int64.compare fd_offset Int64.zero < 0 then
+      efail "pread" EINVAL
+    else
+      let dim = Bigarray.Array1.dim buf in
+      iread ~fd_offset ?pos ?len ~dim ~buf t
 
-  let write ?(pos=0) ?len ~dim t ~buf =
+  let pread ?pos ?len t ~fd_offset ~buf =
+    if Int64.compare fd_offset Int64.zero < 0 then
+      efail "pread" EINVAL
+    else
+     let dim = Bytes.length buf in
+     iread ~fd_offset ?pos ?len ~dim ~buf t
+
+  external iwrite:
+    file -> 'a -> int -> int -> int64 -> loop -> Req.t -> int_cb ->
+    Int_result.unit = "uwt_fs_write_byte" "uwt_fs_write_native"
+
+  let iwrite ~fd_offset ?(pos=0) ?len ~dim t ~buf =
     let len =
       match len with
       | None -> dim - pos
@@ -1690,33 +1702,58 @@ module Fs = struct
     if pos < 0 || len < 0 || pos > dim - len then
       Lwt.fail (Invalid_argument "Uwt.Fs.write")
     else
-      Req.qli ~typ ~name:"write" ~param ~f:(write t buf pos len)
+      Req.qli ~typ ~name:"write" ~param ~f:(iwrite t buf pos len fd_offset)
 
   let write_ba ?pos ?len t ~(buf:buf) =
     let dim = Bigarray.Array1.dim buf in
-    write ~dim ?pos ?len t ~buf
+    iwrite ~fd_offset:Int64.minus_one ~dim ?pos ?len t ~buf
 
   let write_string ?pos ?len t ~buf =
     let dim = String.length buf in
-    write ~dim ?pos ?len t ~buf
+    iwrite ~fd_offset:Int64.minus_one ~dim ?pos ?len t ~buf
 
   let write ?pos ?len t ~buf =
     let dim = Bytes.length buf in
-    write ~dim ?pos ?len t ~buf
+    iwrite ~fd_offset:Int64.minus_one ~dim ?pos ?len t ~buf
 
-  external writev:
-    file -> Iovec_write.t array -> Iovec_write.t list ->
+  let iwrite ~fd_offset ?pos ?len ~dim t ~buf =
+    if Int64.compare fd_offset Int64.zero < 0 then
+      efail "pwrite" EINVAL
+    else
+      iwrite ~fd_offset ~dim ?pos ?len t ~buf
+
+  let pwrite_ba ?pos ?len t ~fd_offset ~(buf:buf) =
+    let dim = Bigarray.Array1.dim buf in
+    iwrite ~fd_offset ~dim ?pos ?len t ~buf
+
+  let pwrite_string ?pos ?len t ~fd_offset ~buf =
+    let dim = String.length buf in
+    iwrite ~fd_offset ~dim ?pos ?len t ~buf
+
+  let pwrite ?pos ?len t ~fd_offset ~buf =
+    let dim = Bytes.length buf in
+    iwrite ~fd_offset ~dim ?pos ?len t ~buf
+
+  external iwritev:
+    file -> Iovec_write.t array -> Iovec_write.t list -> int64 ->
     loop -> Req.t -> int_cb ->
     Int_result.unit =
     "uwt_fs_writev_byte" "uwt_fs_writev_native"
 
-  let writev t iol =
+  let iwritev ~fd_offset t iol =
     let open Iovec_write in
     match prep_for_cstub iol with
     | Invalid -> Lwt.fail (Invalid_argument "Uwt.Fs.writev")
     | Empty -> write ~pos:0 ~len:0 t ~buf:(Bytes.create 1)
     | All_ba(ar,bl) ->
-      Req.qli ~typ ~name:"writev" ~param ~f:(writev t ar bl)
+      Req.qli ~typ ~name:"writev" ~param ~f:(iwritev t ar bl fd_offset)
+
+  let writev a b = iwritev ~fd_offset:Int64.minus_one a b
+  let pwritev a b fd_offset =
+    if Int64.compare fd_offset Int64.zero < 0 then
+      efail "writev" EINVAL
+    else
+      iwritev ~fd_offset a b
 
   external close:
     file -> loop -> Req.t -> unit_cb -> Int_result.unit =